I've upgraded my json views.
Before, it displayed ids of foreign keys. Now, it displays the whole object for each foreign key. For example, it now displays the whole user's information instead of just his id when displaying an article.
For this to take place, I've coded relationships between the 3 models. (It even shows the comments of an article when accessing that article!)